+++ src/main/java/top/rslly/iot/services/ProductServiceImpl.java
@@ -20,23 +20,20 @@
 package top.rslly.iot.services;
 
 import org.springframework.beans.BeanUtils;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import top.rslly.iot.dao.MqttUserRepository;
-import top.rslly.iot.dao.ProductModelRepository;
-import top.rslly.iot.dao.ProductRepository;
-import top.rslly.iot.dao.WxProductBindRepository;
-import top.rslly.iot.models.MqttUserEntity;
-import top.rslly.iot.models.ProductEntity;
-import top.rslly.iot.models.ProductModelEntity;
-import top.rslly.iot.models.WxProductBindEntity;
+import top.rslly.iot.dao.*;
+import top.rslly.iot.models.*;
 import top.rslly.iot.param.request.Product;
+import top.rslly.iot.utility.JwtTokenUtil;
 import top.rslly.iot.utility.result.JsonResult;
 import top.rslly.iot.utility.result.ResultCode;
 import top.rslly.iot.utility.result.ResultTool;
 
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -48,11 +45,54 @@ public class ProductServiceImpl implements ProductService {
   @Resource
   private WxProductBindRepository wxProductBindRepository;
   @Resource
+  private UserProductBindRepository userProductBindRepository;
+  @Resource
   private MqttUserRepository mqttUserRepository;
+  @Resource
+  private WxUserRepository wxUserRepository;
+  @Resource
+  private UserRepository userRepository;
+  @Autowired
+  private SafetyService safetyService;
 
   @Override
-  public JsonResult<?> getProduct() {
-    var result = productRepository.findAll();
+  public JsonResult<?> getProduct(String token) {
+    String token_deal = token.replace(JwtTokenUtil.TOKEN_PREFIX, "");
+    String role = JwtTokenUtil.getUserRole(token_deal);
+    String username = JwtTokenUtil.getUsername(token_deal);
+    List<ProductEntity> result;
+    if (role.equals("ROLE_" + "wx_user")) {
+      if (wxUserRepository.findAllByName(username).isEmpty()) {
+        return ResultTool.fail(ResultCode.COMMON_FAIL);
+      }
+      String openid = wxUserRepository.findAllByName(username).get(0).getOpenid();
+      result = new ArrayList<>();
+      var wxBindProductResponseList = wxProductBindRepository.findProductIdByOpenid(openid);
+      if (wxBindProductResponseList.isEmpty()) {
+        return ResultTool.fail(ResultCode.COMMON_FAIL);
+      }
+      for (var s : wxBindProductResponseList) {
+        List<ProductEntity> productEntities = productRepository.findAllById(s.getProductId());
+        result.addAll(productEntities);
+      }
+    } else if (!role.equals("[ROLE_admin]")) {
+      var userList = userRepository.findAllByUsername(username);
+      if (userList.isEmpty()) {
+        return ResultTool.fail(ResultCode.COMMON_FAIL);
+      }
+      int userId = userList.get(0).getId();
+      result = new ArrayList<>();
+      var userProductBindEntityList = userProductBindRepository.findAllByUserId(userId);
+      if (userProductBindEntityList.isEmpty()) {
+        return ResultTool.fail(ResultCode.COMMON_FAIL);
+      }
+      for (var s : userProductBindEntityList) {
+        List<ProductEntity> productEntities = productRepository.findAllById(s.getProductId());
+        result.addAll(productEntities);
+      }
+    } else {
+      result = productRepository.findAll();
+    }
     if (result.isEmpty()) {
       return ResultTool.fail(ResultCode.COMMON_FAIL);
     } else
@@ -80,7 +120,10 @@ public class ProductServiceImpl implements ProductService {
     List<ProductModelEntity> productModelEntityList = productModelRepository.findAllByProductId(id);
     List<WxProductBindEntity> wxProductBindEntityList =
         wxProductBindRepository.findAllByProductId(id);
-    if (productModelEntityList.isEmpty() && wxProductBindEntityList.isEmpty()) {
+    List<UserProductBindEntity> userProductBindEntityList =
+        userProductBindRepository.findAllByProductId(id);
+    if (productModelEntityList.isEmpty() && wxProductBindEntityList.isEmpty()
+        && userProductBindEntityList.isEmpty()) {
       List<ProductEntity> result = productRepository.deleteById(id);
       if (result.isEmpty())
         return ResultTool.fail(ResultCode.PARAM_NOT_VALID);
